+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Full Keycloak upstream jira issue can be shown if filtered by Fix version.
 
 Our Keycloak version is working well with PostgreSQL database. For using other SQL databases, text field in database need to be evaluated.
 
+## [Unreleased]
+### Fixed
+- Handling of changed profile information during sign-up when review profile is set to "missing" or T&C is requested
+
 ## [22.0.11-1.11] - 2024-11-19
 
 ### Added